# Adding IQ1_TN - 1.6875 bpw for TriLM ternary models